replace qbits from itrex to ipex woq linear
