+++ awq/modules/linear/__init__.py
@@ -1,7 +1,7 @@
 from .exllama import WQLinear_Exllama, exllama_post_init
 from .exllamav2 import WQLinear_ExllamaV2, exllamav2_post_init
 from .gemm import WQLinear_GEMM
-from .gemm_qbits import WQLinear_QBits, qbits_post_init
+from .gemm_ipex import WQLinear_IPEX, ipex_post_init
 from .gemv import WQLinear_GEMV
 from .marlin import WQLinear_Marlin, marlin_post_init
 from .gemv_fast import WQLinear_GEMVFast
