+++ awq/utils/fused_utils.py
@@ -7,7 +7,7 @@ from awq.modules.linear import (
     WQLinear_Exllama,
     WQLinear_ExllamaV2,
     WQLinear_GEMVFast,
-    WQLinear_QBits,
+    WQLinear_IPEX,
 )
 
 
@@ -79,10 +79,10 @@ def fuse_qkv(module, q_proj, k_proj, v_proj):
         q_linear = WQLinear_Marlin
     elif isinstance(q_proj, WQLinear_GEMVFast):
         q_linear = WQLinear_GEMVFast
-    elif isinstance(q_proj, WQLinear_QBits):
-        q_linear = WQLinear_QBits
+    elif isinstance(q_proj, WQLinear_IPEX):
+        q_linear = WQLinear_IPEX
 
-    if isinstance(q_proj, WQLinear_QBits):
+    if isinstance(q_proj, WQLinear_IPEX):
         qkv_layer = q_linear(
             q_proj.w_bit,
             q_proj.group_size,
@@ -113,7 +113,7 @@ def fuse_qkv(module, q_proj, k_proj, v_proj):
             [q_proj.scales, k_proj.scales, v_proj.scales], dim=0
         )
         qkv_layer.split_k_iters = q_proj.split_k_iters
-    elif isinstance(q_proj, WQLinear_GEMM) or isinstance(q_proj, WQLinear_QBits):
+    elif isinstance(q_proj, WQLinear_GEMM) or isinstance(q_proj, WQLinear_IPEX):
         qkv_layer.qweight = torch.cat(
             [q_proj.qweight, k_proj.qweight, v_proj.qweight], dim=1
         )
